Split up files with more than one class so that each class has their own file as per PSR4 standard.
