+++ src/Attachments.php
@@ -1,0 +1,5 @@
+<?php namespace Locker\XApi;
+
+class Attachments extends Collection {
+  protected $member_type = 'Locker\XApi\Attachment';
+}
