+++ src/Boolean.php
@@ -1,0 +1,5 @@
+<?php namespace Locker\XApi;
+
+class Boolean extends TypedAtom {
+  protected static $expected_types = ['boolean'];
+}
