+++ src/Context.php
@@ -13,23 +13,3 @@ class Context extends Element {
     'extensions' => 'Locker\XApi\Extensions'
   ];
 }
-
-class ContextActivities extends Element {
-  protected $props = [
-    'parent' => 'Locker\XApi\Activities',
-    'grouping' => 'Locker\XApi\Activities',
-    'category' => 'Locker\XApi\Activities',
-    'other' => 'Locker\XApi\Activities'
-  ];
-}
-
-class Activities extends Collection {
-  protected $member_type = 'Locker\XApi\Activity';
-
-  public function setValue($new_value) {
-    if (is_object($new_value)) {
-      $new_value = [$new_value];
-    }
-    return parent::setValue($new_value);
-  }
-}
