+++ src/ContextActivities.php
@@ -1,0 +1,10 @@
+<?php namespace Locker\XApi;
+
+class ContextActivities extends Element {
+  protected $props = [
+    'parent' => 'Locker\XApi\Activities',
+    'grouping' => 'Locker\XApi\Activities',
+    'category' => 'Locker\XApi\Activities',
+    'other' => 'Locker\XApi\Activities'
+  ];
+}
