+++ src/Duration.php
@@ -1,0 +1,5 @@
+<?php namespace Locker\XApi;
+
+class Duration extends RegexpAtom {
+  protected static $pattern = '/^P((\d+([\.,]\d+)?Y)?(\d+([\.,]\d+)?M)?(\d+([\.,]\d+)?W)?(\d+([\.,]\d+)?D)?)?(T(\d+([\.,]\d+)?H)?(\d+([\.,]\d+)?M)?(\d+([\.,]\d+)?S)?)?$/i';
+}
