+++ src/Integer.php
@@ -1,0 +1,5 @@
+<?php namespace Locker\XApi;
+
+class Integer extends TypedAtom {
+  protected static $expected_types = ['integer'];
+}
