+++ src/InteractionComponent.php
@@ -1,0 +1,8 @@
+<?php namespace Locker\XApi;
+
+class InteractionComponent extends Element {
+  protected $props = [
+    'id' => 'Locker\XApi\String',
+    'description' => 'Locker\XApi\LanguageMap'
+  ];
+}
