+++ src/Number.php
@@ -1,0 +1,5 @@
+<?php namespace Locker\XApi;
+
+class Number extends TypedAtom {
+  protected static $expected_types = ['integer', 'double'];
+}
