+++ src/Strings.php
@@ -1,0 +1,5 @@
+<?php namespace Locker\XApi;
+
+class Strings extends Collection {
+  protected $member_type = 'Locker\XApi\String';
+}
